requirements.txt
aiogram>=3.6.0
aiohttp>=3.9.5
aiosqlite>=0.20.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.8
aioresponses>=0.7.6
loguru>=0.7.2
base58>=2.1.1
